pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fidelity"
version = "0.1.0"
description = "Fine-grained Interpretable Distillation for Effective Language Insights and Topic Yielding"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "python-dotenv",
    "numpy",
    "pandas",
    "tqdm",
    "torch",
    "scikit-learn",
    "hdbscan",
    "umap-learn",
    "keybert",
    "gensim",
    "spacy",
    "nltk",
    "sentence-transformers",
]

[tool.setuptools.packages.find]
include = ["fidelity*"]
